# PERF: faster searchsorted in tzconversion