```
ts = pd.Timestamp(0)
ts2 = ts.tz_localize("US/Pacific")

%timeit ts.tz_localize("US/Pacific")
35.1 µs ± 385 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- main
16.8 µs ± 312 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <- PR

%timeit ts2.tz_localize(None)
5.1 µs ± 54.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <- main
2.7 µs ± 68.4 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <- PR
```